clients/demo/requirements.txt
agency-swarm>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0 